local_qna_chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
numpy>=2.0.0
pandas>=2.2.0
faiss-cpu
sentence-transformers>=2.2.2
ollama==0.1.5
python-multipart==0.0.6
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pypdf==3.17.1
requests==2.31.0
